refactor: move decodeRevertReason to contracts/utils and use in error handling
