+++ packages/sdk/src/thor/thor-client/contracts/index.ts
@@ -1,6 +1,7 @@
 export { ContractsModule } from './contracts-module';
 export * from './model';
 export * from './types';
+export { decodeRevertReason } from './utils';
 // Viem compatibility exports
 export { createViemContract, getContract } from './viem-adapter';
 export type { ViemContract } from './viem-adapter';
